+++ include/nstd/audio.h
@@ -171,12 +171,12 @@ NSTDAPI NSTDAudioSink nstd_audio_sink_new(const NSTDAudioPlayStream *const strea
 /// Parameters:
 ///     `NSTDAudioSink sink` - The audio sink.
 ///     `const NSTDFile *const file` - The audio file.
-///     `const int should_loop` - Nonzero if the audio should be looped.
-/// Returns: `int errc` - Nonzero on error.
-NSTDAPI int nstd_audio_sink_append_from_file(
+///     `const NSTDBool should_loop` - Nonzero if the audio should be looped.
+/// Returns: `NSTDErrorCode errc` - Nonzero on error.
+NSTDAPI NSTDErrorCode nstd_audio_sink_append_from_file(
     NSTDAudioSink sink,
     const NSTDFile *const file,
-    const int should_loop);
+    const NSTDBool should_loop);
 
 /// Plays an audio sink.
 /// Parameters: