+++ src/audio.rs
@@ -1,5 +1,5 @@
 use crate::{
-    core::def::{NSTDAny, NSTDAnyConst},
+    core::def::{NSTDAny, NSTDAnyConst, NSTDBool, NSTDErrorCode},
     fs::file::NSTDFile,
     string::NSTDString,
 };
@@ -305,17 +305,17 @@ pub unsafe extern "C" fn nstd_audio_sink_new(stream: &NSTDAudioPlayStream) -> NS
 /// Parameters:
 ///     `NSTDAudioSink sink` - The audio sink.
 ///     `const NSTDFile *const file` - The audio file.
-///     `const int should_loop` - Nonzero if the audio should be looped.
-/// Returns: `int errc` - Nonzero on error.
+///     `const NSTDBool should_loop` - Nonzero if the audio should be looped.
+/// Returns: `NSTDErrorCode errc` - Nonzero on error.
 #[cfg_attr(feature = "clib", no_mangle)]
 pub unsafe extern "C" fn nstd_audio_sink_append_from_file(
     sink: NSTDAudioSink,
     file: &NSTDFile,
-    should_loop: c_int,
-) -> c_int {
+    should_loop: NSTDBool,
+) -> NSTDErrorCode {
     let buf = BufReader::new((&*file.handle).get_ref());
     match should_loop {
-        0 => match Decoder::new(buf) {
+        NSTDBool::NSTD_BOOL_FALSE => match Decoder::new(buf) {
             Ok(decoder) => {
                 (*sink).append(decoder);
                 0